+++ CloudShell.Abstractions/ResourceManager/Resource.cs
@@ -226,6 +226,7 @@ public static class ResourceAttributeNames
     public const string DeploymentRevision = "deployment.revision";
     public const string DeploymentWorkloadVersion = "deployment.workloadVersion";
     public const string DeploymentDesiredReplicas = "deployment.replicas.desired";
+    public const string DeploymentMaterializedReplicas = "deployment.replicas.materialized";
     public const string DeploymentProjectedReplicas = "deployment.replicas.projected";
     public const string RuntimeKind = "runtime.kind";
     public const string RuntimeContainerName = "runtime.container.name";